fix : ChampionshipForm 대회 생성 수정후 이동 버그 수정했습니다
